Updated Code as per Requirements and Standards
